Back to WAMR on all platforms
Runtime was not designed to work in plugin environments. Add a patch for
the C API that avoids the usage of a singleton for engine instances, and
remove special handling of WAMR from WasmEngine.cpp.

This patch plus configuring the runtime with WAMR_DISABLE_HW_BOUND_CHECK=1
allows it to work on all platforms. If new problems arise in the future
there is always Wasmer for Linux and macOS. And Windows does not suffer
too much from WAMR design because the way static data works for DLLs.

So far the ideal solution seems to be Wasmer for all platforms but it is
still crashing on some Windows DAWs including popular Ableton Live.
Need to ask Wasmer team if their library is really safe to initialize
multiple times...