Fix telecomp cropped UI on Windows/REAPER/HiDPI
For some reason REAPER is providing a window with ~90% of the expected
area when display scaling is set to 150%. Ableton Live does not exhibit
this behavior. It causes the web UI to appear cropped when setting
document.body width/height to fixed px values.

Use 100% width/height when running the UI in the plugin embedded web
view until the root cause for the missing area in REAPER (and
potentially other hosts) is found.