BSONVariant from/to JSON text
This commit makes cJSON-based JSONVariant redundant. It is
still preferable when the messaging protocol is text-based
(the default), because it does not require building a full
library like libbson. cJSON is implemented in a single
source file + header, therefore it is already included in
this repository. JSONVariant implementation is also less
complex because cJSON instances can wrap a single scalar
value, while libbson only allows "documents" (key/value
collections). Arrays are documents with numeric keys.