More self-explanatory API for the shared memory
DPF hard Plugin/UI decoupling plus the fact state notifications
are unidirectional (UI->Plugin) for the moment (updateStateValue()
is under construction), forces the UI to own the shared memory
instead of Plugin. The inverse sounds like the natural choice and
it would be in fact simpler to implement, because shmem lifetime
would be equal to Plugin lifetime.

Since now shmem lifetime is equal to the UI lifetime, some caution
is needed to make sure things do not go wild in the audio thread
when accessing shared stuff. Visualization data is a good example.

There is always DISTRHO_PLUGIN_WANT_DIRECT_ACCESS ...