Wave example visualization backend
Beam 8-bit samples to network clients. Use local transport for
plug-in embedded web view, and binary WebSocket for network
clients.

TCP+WS is not ideal for streaming audio, the purpose of this demo
is not streaming audio but showing how to transfer complex data
from the Plugin instance to the UI instance using shared memory,
and then how to transmit such data to external clients.

- For real world low-latency audio streaming UDP is preferred

- Visualization data could be rendered on the native side then
  streamed to clients for lower network traffic

- DISTRHO_PLUGIN_WANT_DIRECT_ACCESS could be enabled for a simpler
  design without shared memory