Bugfix: shmem disconnected after changing programs
Changing programs does too much automatic stuff
on the UI instance, eg. calls stateChanged() and
probably more.

Be defensive and try to detect bogus stateChanged()
calls (ie. when state has not really changed, to
avoid doing things twice that are not meant to be
done twice).

Probably all code from now on should be more
defensive and not rely much on the DPF Plugin and
UI lifetime, it is not deeply documented after all.